+++ packages/memory-honcho/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @theokit/sdk@4.0.0
+
 ## 2.0.0
 
 ### Patch Changes